requirements.txt
torch==2.0.1
shap==0.44.1
scikit-learn==1.5.0
pefile==2023.2.7
numpy==1.26.4
lime==0.2.0.1
lief==0.13.0
interpret-community==0.32.0
captum==0.7.0
click~=8.1.8
pandas~=2.0.0
joblib==1.4.2
scipy==1.13.1
matplotlib==3.7.1
seaborn==0.13.2
pyelftools==0.31
pyarrow==20.0.0
ipywidgets~=8.1.7
